:book: Manpage tweaks
Long story short, the priority at this time is to get a manual page
generated dynamically from git_guilt.guilt using the sphinx-argparse
extension. To this end, I'm happy to have the 'argparse' block in the
manpage rST include ":manpage: True" even though it *does* mean the page
looks a bit wonky when rendered to HTML on readthedocs.org.

There is a bug in sphinx-argparse 0.1.14 that causes the
manpage-specific docutils builder to create a malformed DOM. I've
submitted a fix upstream that should take care of this.